requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
cython==0.29.37
    # via hdbscan
filelock==3.14.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
    #   triton
fsspec==2024.3.1
    # via
    #   huggingface-hub
    #   torch
hdbscan==0.8.33
    # via -r requirements.in
huggingface-hub==0.22.2
    # via
    #   sentence-transformers
    #   tokenizers
    #   transformers
idna==3.7
    # via requests
jinja2==3.1.3
    # via torch
joblib==1.4.0
    # via
    #   hdbscan
    #   pynndescent
    #   scikit-learn
llvmlite==0.42.0
    # via
    #   numba
    #   pynndescent
markupsafe==2.1.5
    # via jinja2
mpmath==1.3.0
    # via sympy
networkx==3.3
    # via torch
numba==0.59.1
    # via
    #   pynndescent
    #   umap-learn
numpy==1.26.4
    # via
    #   hdbscan
    #   numba
    #   scikit-learn
    #   scipy
    #   sentence-transformers
    #   transformers
    #   umap-learn
nvidia-cublas-cu12==12.1.3.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.1.105
    # via torch
nvidia-cuda-nvrtc-cu12==12.1.105
    # via torch
nvidia-cuda-runtime-cu12==12.1.105
    # via torch
nvidia-cudnn-cu12==8.9.2.26
    # via torch
nvidia-cufft-cu12==11.0.2.54
    # via torch
nvidia-curand-cu12==10.3.2.106
    # via torch
nvidia-cusolver-cu12==11.4.5.107
    # via torch
nvidia-cusparse-cu12==12.1.0.106
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-nccl-cu12==2.20.5
    # via torch
nvidia-nvjitlink-cu12==12.4.127
    # via
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
nvidia-nvtx-cu12==12.1.105
    # via torch
packaging==24.0
    # via
    #   huggingface-hub
    #   transformers
pillow==10.3.0
    # via sentence-transformers
pynndescent==0.5.12
    # via umap-learn
pyyaml==6.0.1
    # via
    #   huggingface-hub
    #   transformers
regex==2024.4.28
    # via transformers
requests==2.31.0
    # via
    #   huggingface-hub
    #   transformers
safetensors==0.4.3
    # via transformers
scikit-learn==1.4.2
    # via
    #   hdbscan
    #   pynndescent
    #   sentence-transformers
    #   umap-learn
scipy==1.13.0
    # via
    #   hdbscan
    #   pynndescent
    #   scikit-learn
    #   sentence-transformers
    #   umap-learn
sentence-transformers==2.7.0
    # via -r requirements.in
sympy==1.12
    # via torch
threadpoolctl==3.5.0
    # via scikit-learn
tokenizers==0.19.1
    # via transformers
torch==2.3.0
    # via sentence-transformers
tqdm==4.66.2
    # via
    #   huggingface-hub
    #   sentence-transformers
    #   transformers
    #   umap-learn
transformers==4.40.1
    # via sentence-transformers
triton==2.3.0
    # via torch
typing-extensions==4.11.0
    # via
    #   huggingface-hub
    #   torch
umap-learn==0.5.6
    # via -r requirements.in
urllib3==2.2.1
    # via requests
